Move shared hooks from utils under hooks folder
