+++ app/src/auth/components/LogoutPage/index.jsx
@@ -5,7 +5,7 @@ import { connect } from 'react-redux'
 
 import { post } from '$shared/utils/api'
 import useIsMounted from '$shared/hooks/useIsMounted'
-import useOnMount from '$shared/utils/useOnMount'
+import useOnMount from '$shared/hooks/useOnMount'
 import { logout as logoutAction } from '$shared/modules/user/actions'
 import ErrorPageView from '$mp/components/ErrorPageView'
 import routes from '$routes'
