+++ app/src/auth/components/RegisterPage/index.jsx
@@ -6,8 +6,8 @@ import * as yup from 'yup'
 import { I18n, Translate } from 'react-redux-i18n'
 import { userIsNotAuthenticated } from '$mp/utils/auth'
 
-import useIsMountedRef from '$shared/utils/useIsMountedRef'
-import useOnMount from '$shared/utils/useOnMount'
+import useIsMountedRef from '$shared/hooks/useIsMountedRef'
+import useOnMount from '$shared/hooks/useOnMount'
 import AuthFormProvider from '../AuthFormProvider'
 import SessionProvider from '../SessionProvider'
 import AuthFormContext from '../../contexts/AuthForm'
