+++ app/src/auth/components/ResetPasswordPage/index.jsx
@@ -7,8 +7,8 @@ import qs from 'query-string'
 import { I18n, Translate } from 'react-redux-i18n'
 import { push } from 'connected-react-router'
 
-import useIsMountedRef from '$shared/utils/useIsMountedRef'
-import useOnMount from '$shared/utils/useOnMount'
+import useIsMountedRef from '$shared/hooks/useIsMountedRef'
+import useOnMount from '$shared/hooks/useOnMount'
 import AuthFormProvider from '../AuthFormProvider'
 import { userIsNotAuthenticated } from '$mp/utils/auth'
 import AuthFormContext from '$auth/contexts/AuthForm'
