+++ app/src/editor/canvas/components/CanvasController/Run.jsx
@@ -5,7 +5,7 @@
 import React, { useContext, useState, useCallback, useMemo } from 'react'
 import get from 'lodash/get'
 
-import useIsMountedRef from '$shared/utils/useIsMountedRef'
+import useIsMountedRef from '$shared/hooks/useIsMountedRef'
 import * as SubscriptionStatus from '$editor/shared/components/SubscriptionStatus'
 import { Context as PermissionContext } from '$editor/canvas/hooks/useCanvasPermissions'
 import usePending from '$shared/hooks/usePending'
