+++ app/src/editor/canvas/components/CanvasController/useCanvasCreateCallback.jsx
@@ -1,6 +1,6 @@
 import { useContext, useCallback } from 'react'
 
-import useIsMountedRef from '$shared/utils/useIsMountedRef'
+import useIsMountedRef from '$shared/hooks/useIsMountedRef'
 import { Context as RouterContext } from '$shared/components/RouterContextProvider'
 import usePending from '$shared/hooks/usePending'
