+++ app/src/editor/canvas/components/CanvasController/useCanvasUpdater.jsx
@@ -1,6 +1,6 @@
 import { useMemo, useContext } from 'react'
 
-import useMountedCallback from '$shared/utils/useMountedCallback'
+import useOnlyIfMountedCallback from '$shared/hooks/useOnlyIfMountedCallback'
 import { Context as UndoContext } from '$shared/components/UndoContextProvider'
 
 import * as CanvasState from '../../state'
@@ -16,11 +16,11 @@ function canvasUpdater(fn) {
 export default function useCanvasUpdater() {
     const { push, replace } = useContext(UndoContext)
 
-    const setCanvas = useMountedCallback((action, fn, done) => {
+    const setCanvas = useOnlyIfMountedCallback((action, fn, done) => {
         push(action, canvasUpdater(fn), done)
     }, [push, canvasUpdater])
 
-    const replaceCanvas = useMountedCallback((fn, done) => {
+    const replaceCanvas = useOnlyIfMountedCallback((fn, done) => {
         replace(canvasUpdater(fn), done)
     }, [push, canvasUpdater])
